wasi: add simple wasi test/example
This commit adds a very simply wasi example that calls a _start
function on a WebAssembly module that has been compiled with a
WASI compatibly compiler.

This is the source c program:

void print() {
  printf("simple-wasi...\n");
}

int main(int argc, char** argv) {
  print();
  return 0;
}

And it can be run with wasmtime:
$ wasmtime simple-wasi.wasm
simple-wasi...

The test test-wasi-binding.js currently does not print anything as these
functions (WASI::FdPrestatGet etc) are not implemented yet. But I think
it still might be useful to have an example/test that shows how this can
be used.