wasi: sync with latest uvwasi
This update gets uvwasi + Node building on all of the platforms
that Node runs in the CI (instead of just Ubuntu, macOS, and
Windows that are offered in GitHub's CI). It refactors file
detection logic, improves the handling of clocks, fixes some
incorrect constant values, and a few other small bug fixes.